Integrate custom import package name checker into the roxvet framework.
