+++ tools/roxvet/roxvet.go
@@ -3,6 +3,7 @@ package main
 import (
 	"github.com/stackrox/rox/tools/roxvet/analyzers/dontprintferr"
 	"github.com/stackrox/rox/tools/roxvet/analyzers/filepathwalk"
+	"github.com/stackrox/rox/tools/roxvet/analyzers/importpackagenames"
 	"github.com/stackrox/rox/tools/roxvet/analyzers/invalidoutputroxctl"
 	"github.com/stackrox/rox/tools/roxvet/analyzers/needlessformat"
 	"github.com/stackrox/rox/tools/roxvet/analyzers/protoclone"
@@ -28,5 +29,6 @@ func main() {
 		invalidoutputroxctl.Analyzer,
 		filepathwalk.Analyzer,
 		validateimports.Analyzer,
+		importpackagenames.Analyzer,
 	)
 }
